+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 * Eliminated dependency in Magento\Data\Form from Magento\Core module
 * Eliminated Magento\Media module
 * Fixed copyright information
+* Removed page type hierarchy
+* Removed page fragment types
 
 2.0.0.0-dev49
 =============